Move all ssh commands to a subcategory in user...
